+++ 3715/1/gemlist.rb
@@ -1,0 +1,9 @@
+require_relative 'file_interaction.rb'
+require_relative 'rubygem.rb'
+
+# Making a list of gems
+class GemList
+  def self.make_gemlist
+    Files.new.name_gem.map { |title| RubyGem.new(title) }
+  end
+end
